Fixed several unit-tests
